+++ main.go
@@ -5,6 +5,8 @@ import (
 	"net/http"
     "os"
     "github.com/fujiwara-labo/laboratory-assignment.git/crypto"
+    "github.com/fujiwara-labo/laboratory-assignment.git/models"
+    "github.com/fujiwara-labo/laboratory-assignment.git/control"
     _ "github.com/go-sql-driver/mysql"
 
     "github.com/gin-gonic/gin"
@@ -14,64 +16,64 @@ import (
 )
 
 // User モデルの宣言
-type User struct {
-	gorm.Model
-	Username string `form:"username" binding:"required" gorm:"unique;not null"`
-	Password string `form:"password" binding:"required"`
-}
+// type User struct {
+// 	gorm.Model
+// 	Username string `form:"username" binding:"required" gorm:"unique;not null"`
+// 	Password string `form:"password" binding:"required"`
+// }
 
-func gormConnect() *gorm.DB {
-	err := godotenv.Load()
-	if err != nil {
-		log.Fatal(err)
-	}
-    DBMS := os.Getenv("DRIVER")
-    CONNECT := os.Getenv("DSN")
-	// USER := os.Getenv("MYSQL_USER")
-	// PASS := os.Getenv("MYSQL_PASSWORD")
-	// DBNAME := os.Getenv("MYSQL_DATABASE")
-	// CONNECT := USER + ":" + PASS + "@/" + DBNAME + "?parseTime=true"
-	db, err := gorm.Open(DBMS, CONNECT)
-	if err != nil {
-		panic(err.Error())
-	}
-	return db
-}
-// DBの初期化
-func dbInit() {
-	db := gormConnect()
-	// コネクション解放
-	defer db.Close()
-    //構造体に基づいてテーブルを作成
-	db.AutoMigrate(&User{})
-}
+// func gormConnect() *gorm.DB {
+// 	err := godotenv.Load()
+// 	if err != nil {
+// 		log.Fatal(err)
+// 	}
+//     DBMS := os.Getenv("DRIVER")
+//     CONNECT := os.Getenv("DSN")
+// 	// USER := os.Getenv("MYSQL_USER")
+// 	// PASS := os.Getenv("MYSQL_PASSWORD")
+// 	// DBNAME := os.Getenv("MYSQL_DATABASE")
+// 	// CONNECT := USER + ":" + PASS + "@/" + DBNAME + "?parseTime=true"
+// 	db, err := gorm.Open(DBMS, CONNECT)
+// 	if err != nil {
+// 		panic(err.Error())
+// 	}
+// 	return db
+// }
+// // DBの初期化
+// func dbInit() {
+// 	db := gormConnect()
+// 	// コネクション解放
+// 	defer db.Close()
+//     //構造体に基づいてテーブルを作成
+// 	db.AutoMigrate(&User{})
+// }
 
 // ユーザー登録処理
-func createUser(username string, password string) []error {
-	passwordEncrypt, _ := crypto.PasswordEncrypt(password)
-	db := gormConnect()
-	defer db.Close()
-	// Insert処理
-	if err := db.Create(&User{Username: username, Password: passwordEncrypt}).GetErrors(); err != nil {
-		return err
-	}
-	return nil
+// func createUser(username string, password string) []error {
+// 	passwordEncrypt, _ := crypto.PasswordEncrypt(password)
+// 	db := gormConnect()
+// 	defer db.Close()
+// 	// Insert処理
+// 	if err := db.Create(&User{Username: username, Password: passwordEncrypt}).GetErrors(); err != nil {
+// 		return err
+// 	}
+// 	return nil
 
-}
-// ユーザーを一件取得
-func getUser(username string) User {
-	db := gormConnect()
-	var user User
-	db.First(&user, "username = ?", username)
-	db.Close()
-	return user
-}
+// }
+// // ユーザーを一件取得
+// func getUser(username string) User {
+// 	db := gormConnect()
+// 	var user User
+// 	db.First(&user, "username = ?", username)
+// 	db.Close()
+// 	return user
+// }
 
 func main() {
     router := gin.Default()
 	router.LoadHTMLGlob("views/*.html")
 
-    dbInit()
+    control.dbInit()
 
     router.GET("/", func(c *gin.Context) {
         c.HTML(http.StatusOK, "index.html", gin.H{
@@ -88,7 +90,7 @@ func main() {
 
     // ユーザー登録
     router.POST("/signup", func(c *gin.Context) {
-        var form User
+        var form models.User
         // バリデーション処理
         if err := c.Bind(&form); err != nil {
             c.HTML(http.StatusBadRequest, "signup.html", gin.H{"err": err})
@@ -97,7 +99,7 @@ func main() {
             username := c.PostForm("username")
             password := c.PostForm("password")
             // 登録ユーザーが重複していた場合にはじく処理
-            if err := createUser(username, password); err != nil {
+            if err := control.createUser(username, password); err != nil {
                 c.HTML(http.StatusBadRequest, "signup.html", gin.H{"err": err})
             }
             c.Redirect(302, "/")
@@ -114,7 +116,7 @@ func main() {
     router.POST("/login", func(c *gin.Context) {
 
         // DBから取得したユーザーパスワード(Hash)
-        dbPassword := getUser(c.PostForm("username")).Password
+        dbPassword := control.getUser(c.PostForm("username")).Password
         log.Println(dbPassword)
         // フォームから取得したユーザーパスワード
         formPassword := c.PostForm("password")
