+++ models/user.go
@@ -1,0 +1,11 @@
+package models
+
+import(
+	"github.com/jinzhu/gorm"
+)
+// User モデルの宣言
+type User struct {
+	gorm.Model
+	Username string `form:"username" binding:"required" gorm:"unique;not null"`
+	Password string `form:"password" binding:"required"`
+}
